+++ zone/zonedump.h
@@ -99,8 +99,8 @@ struct NPCType
 	int             charm_atk;
 	int16           attack_count;
 	char            special_abilities[512];
-	uint16          d_melee_texture1;
-	uint16          d_melee_texture2;
+	uint32          d_melee_texture1;
+	uint32          d_melee_texture2;
 	char            ammo_idfile[30];
 	uint8           prim_melee_type;
 	uint8           sec_melee_type;
